# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shopecart E-commerce API Documentation",
        "description": "Documentation des points d'accès (endpoints) de l'API Shopecart.",
        "contact": {
            "email": "contact@shopecart.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de l'API locale"
        }
    ],
    "paths": {
...
        "/api/v1/notifications/test-registration": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Teste l'envoi de l'e-mail de confirmation d'inscription.",
                "description": "Nécessite qu'un utilisateur avec l'ID 1 existe.",
                "operationId": "testRegistrationEmail",
                "responses": {
                    "200": {
                        "description": "E-mail de test d'inscription envoyé avec succès (Vérifiez la console pour les erreurs)."
                    },
                    "404": {
                        "description": "Utilisateur non trouvé."
                    },
                    "500": {
                        "description": "Erreur d'envoi d'e-mail."
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrez le jeton Bearer obtenu après la connexion.",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}